+++ ide-common/src/main/kotlin/org/digma/intellij/plugin/progress/ProgressUtils.kt
@@ -1,0 +1,9 @@
+package org.digma.intellij.plugin.progress
+
+import com.intellij.openapi.progress.ProgressManager
+
+
+fun runBackgroundTaskInProgressWithRetry(task: RetryableTask) {
+    val backgroundTask = RetryableBackgroundTask(task)
+    ProgressManager.getInstance().run(backgroundTask)
+}